+++ core/bot_manager.py
@@ -1,6 +1,7 @@
 from commons.custom_logger import CustomLogger
 from models.bot_config import BotConfig
 import core.bot_config_loader as bot_config_loader
+from core.bot import Bot
 
 BOT_CONFIG_PATH = 'config/bots_config.json'
 
@@ -8,6 +9,7 @@ class BotManager:
     def __init__(self):
         self.logger = CustomLogger(name=self.__class__.__name__)
         self.bot_config_loader = bot_config_loader
+
 
     def _load_bots_config(self):
         return self.bot_config_loader.load_config(file_path=BOT_CONFIG_PATH)
@@ -20,40 +22,20 @@ class BotManager:
 
         for bot_config in self.bots_config:
             try:
-                self.logger.debug(bot_config)
+                self.logger.debug(message=bot_config)
+                self.logger.info(message=f'Loading {bot_config.bot_name}')
+                bot: Bot = Bot(bot_config=bot_config)
 
-                # self.bots.append(bot)
-                # self.bot_runners.append(bot_runner)
             except Exception as e:
-                self.logger.error_e(f"Failed to create bot runner:", e)
+                self.logger.error_e(message=f"Failed to create bot runner:", e=e)
 
-        self.logger.info(f"Loaded  {len(self.bots_config)}  🤖")
+        self.logger.info(message=f"Loaded  {len(self.bots_config)}  🤖")
 
 
-    # def run_bots(self):
-    #     # run in backtest mode
-    #     if self.run_mode == RunMode.BACKTEST:
-    #         for bot_runner in self.bot_runners:
-    #             if bot_runner.activate_bot.mode != RunMode.BACKTEST:
-    #                 continue
-    #             self.logger.info(f'Runnig  🤖   {bot_runner.bot_fullname}')
-    #             run_dict = bot_runner.run_bot()
-        
-    #             self.runs.append(run_dict)
 
-    #         print_result_table(self.runs)
-
-    #         self.logger.info(f"Total   🤖   run:  {len(self.runs)}")
-    #         self.logger.info(f"Bye!")
-    #         return
-
-    #     # run forwardtest or live mode
-    #     while True:
-    #         for bot_runner in self.bot_runners:
-    #             if bot_runner.activate_bot.mode == RunMode.BACKTEST:
-    #                 continue
-    #             self.logger.info(f'Runnig  🤖   {bot_runner.bot_fullname}')
-    #             run_dict = bot_runner.run_bot()
-    #         self.logger.info(f"Total   🤖   run:  {len(self.runs)}")
+# self.logger.info(f'Runnig  🤖   {bot_runner.bot_fullname}')
+# self.logger.info(f"Total   🤖   run:  {len(self.runs)}")
+# self.logger.info(f'Runnig  🤖   {bot_runner.bot_fullname}')
+# self.logger.info(f"Total   🤖   run:  {len(self.runs)}")
 
 # EOF
