+++ models/enum/entry_strategy.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+class EntryStrategy(Enum):
+    MACDHIST_STATE = "MACDHIST_STATE"
+    MACDHIST_EMA_V1 = "MACDHIST_EMA_V1"
+
+# EOF
