+++ models/enum/exit_strategy.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class ExitStrategy(Enum):
+    MACDHIST_STATE = "MACDHIST_STATE"
+
+# EOF
